components/admin/requirements.txt
Django==2.0
djongo
djangorestframework
django-simple-history==2.3.0